feat(jsonrpc): implement clear signing for the attestation transaction
The remote signer API now supports clear signing. This means that in addition
to the transaction hash, complete transaction data is sent to the remote
signer. We use the JSON-RPC 0.8.1 specification's INVOKE_TXN_V3 schema
to send the transaction data.

To implement clear signing with starknet-rs, we had to duplicate some
code from starknet-rs. Because the transaction data is _not_ available
in the `Signer` trait, we had to implement our own `Account` that has
access to all the details and can then use our own `AttestationSigner`
trait to sign the transaction.