fix(jsonrpc): create JSON-RPC client with a HTTP timeout of 30 seconds
The default HTTP transport in starknet-providers creates a HTTP client with
no timeouts. Since this could block our state machine indefinitely we
now set a timeout of 30 seconds.

Unfortunately starknet-rs uses an older version of reqwest so we have
to add an additional dependency to our Cargo.toml to be able to set
the timeout. This is a temporary workaround until starknet-rs
upgrades to a newer version of reqwest.