feat: retry getting attestation info
There is a use case where stakers would start up the attestation tool
before registering their staker so that they can avoid slashing due to
not submitting an attestation transaction.

This change implements that by retrying getting the initial attestation
info indefinitely. That is, if we fail to fetch the attestation info for
the address configured we just retry in five seconds.