feat: add support for non-ECDSA remote signatures
When using a remote signer, the attestation tool should not force that
the signature returned is an ECDSA signature (`r` and `s` values).

Since on the remote signer API we're expecting the signature as an
array of felts, the attestation tool should just take this array directly
and use it as a signature for the transaction.

Some account contracts _do_ support non-ECDSA signatures, or use a signature
encoding scheme that is not compatible with the ECDSA scheme used by the
OpenZeppelin account contract. (Braavos for example uses a signature
that's longer than just two felts.)