fix(jsonrpc): do transaction submission using pre-confirmed data
We have to do fee estimation on the pre-confirmed block, so that we do
not get an out-of-attestation-window error when doing estimation.

We're doing the submission (and estimation) when the _latest_ block is
still out of the attestation window, but the block that is being built
is within the window. By using the pre-confirmed block for fee estimation,
we can avoid out-of-attestation-window errors.